+++ src/main/java/com/nubble/backend/post/controller/PostApiController.java
@@ -7,6 +7,10 @@ import com.nubble.backend.config.resolver.UserSession;
 import com.nubble.backend.interceptor.session.SessionRequired;
 import com.nubble.backend.post.controller.PostRequest.PostCreateRequest;
 import com.nubble.backend.post.controller.PostResponse.PostCreateResponse;
+import com.nubble.backend.post.controller.mapper.CommentQueryMapper;
+import com.nubble.backend.post.controller.mapper.MemberCommentCommandMapper;
+import com.nubble.backend.post.controller.mapper.PostCommandMapper;
+import com.nubble.backend.post.controller.mapper.PostResponseMapper;
 import com.nubble.backend.post.service.PostCommand.PostCreateCommand;
 import com.nubble.backend.post.service.PostCommand.PostUpdateCommand;
 import com.nubble.backend.post.service.PostService;
