+++ src/main/java/com/nubble/backend/post/controller/mapper/CommentQueryMapper.java
@@ -1,4 +1,4 @@
-package com.nubble.backend.post.controller;
+package com.nubble.backend.post.controller.mapper;
 
 import com.nubble.backend.comment.service.CommentQuery.PostByIdQuery;
 import com.nubble.backend.comment.service.CommentQuery.UserByIdQuery;
