+++ src/main/java/com/nubble/backend/post/controller/mapper/PostCommandMapper.java
@@ -1,4 +1,4 @@
-package com.nubble.backend.post.controller;
+package com.nubble.backend.post.controller.mapper;
 
 import com.nubble.backend.post.controller.PostRequest.PostCreateRequest;
 import com.nubble.backend.post.controller.PostRequest.PostUpdateRequest;
